refactor: move script related code to bitcoin_script module
